Ajustar abordagem de separação de trabalho entre os processos.

+++ src/sequential_Erastotenes.c
@@ -1,6 +1,7 @@
-#include <simple_buffers.h>
-#include <macros.h>
 #include <sequential_Erastotenes.h>
+
+#include <macros.h>
+#include <simple_buffers.h>
 #include <stdint.h>
 #include <inttypes.h>
 #include <string.h>
@@ -20,7 +21,7 @@ u32Buffer sequential_primes_until(u32 const N) {
     LOG("Algorithm called with N < 3");
     return NULL_U32BUFFER;
   }
-  i8Buffer marked_odds = i8_buffer_init((N + 1) / 2 - 1);
+  i8Buffer marked_odds = i8_buffer_init(((u64)N + 1) / 2 - 1);
   if(IS_NULL_BUFFER(marked_odds)) {
     fprintf(stderr, "Failed to alloc memory for the array used to tag prime numbers");
     LOG("Returning null buffer due to memory errors");
@@ -28,7 +29,7 @@ u32Buffer sequential_primes_until(u32 const N) {
   }
   memset(marked_odds.m_buffer, IS_PRIME, marked_odds.m_buffer_size);
   LOG("All values marked as prime with memset");
-  u32Buffer primes = u32_buffer_init((N + 1) / 2); 
+  u32Buffer primes = u32_buffer_init(((u64)N + 1) / 2); 
   if(IS_NULL_BUFFER(primes)) {
     fprintf(stderr, "Failed to alloc memory for the array containing the prime numbers");
     LOG("Returning null buffer due to memory errors");
@@ -43,7 +44,7 @@ u32Buffer sequential_primes_until(u32 const N) {
     u32 const val = num_of(i);
     LOG("Found a prime (number = %" PRIu32 ", index = %" PRIu32 ")", val, i);
     primes.m_buffer[prime_next_index++] = val;
-    for(u32 k = (val * val - 3) / 2; k < marked_odds.m_buffer_size; k += val)
+    for(u64 k = ((u64)val * val - 3) / 2; k < marked_odds.m_buffer_size; k += val)
       marked_odds.m_buffer[k] = NOT_PRIME;
   }
   primes.m_buffer_size = prime_next_index;
